Trekk logikken i jobbene ut i servicer
