+++ service/src/main/kotlin/no/nav/su/se/bakover/service/brev/BrevService.kt
@@ -2,22 +2,19 @@ package no.nav.su.se.bakover.service.brev
 
 import arrow.core.Either
 import no.nav.su.se.bakover.common.persistence.TransactionContext
-import no.nav.su.se.bakover.domain.brev.BrevbestillingId
 import no.nav.su.se.bakover.domain.brev.LagBrevRequest
 import no.nav.su.se.bakover.domain.dokument.Dokument
-import no.nav.su.se.bakover.domain.dokument.Dokumentdistribusjon
-import no.nav.su.se.bakover.domain.journal.JournalpostId
 import java.util.UUID
 
 interface BrevService {
     fun lagBrev(request: LagBrevRequest): Either<KunneIkkeLageBrev, ByteArray>
-    fun distribuerBrev(journalpostId: JournalpostId): Either<KunneIkkeDistribuereBrev, BrevbestillingId>
     fun lagreDokument(dokument: Dokument.MedMetadata)
     fun lagreDokument(dokument: Dokument.MedMetadata, transactionContext: TransactionContext)
-    fun journalførDokument(dokumentdistribusjon: Dokumentdistribusjon): Either<KunneIkkeJournalføreDokument, Dokumentdistribusjon>
-    fun distribuerDokument(dokumentdistribusjon: Dokumentdistribusjon): Either<KunneIkkeBestilleBrevForDokument, Dokumentdistribusjon>
-    fun hentDokumenterForDistribusjon(): List<Dokumentdistribusjon>
     fun hentDokumenterFor(hentDokumenterForIdType: HentDokumenterForIdType): List<Dokument>
+
+    /** Journalfører og distribuerer utgående dokumenter som ikke har blitt journalført / distribuert.
+     * Denne er i utgangspunktet ment å kalles av en kron-jobb. */
+    fun journalførOgDistribuerUtgåendeDokumenter()
 }
 
 /**
